requirements.txt
# Screen Sharing Application Dependencies

# Core Screen Capture
mss==9.0.1
Pillow==10.1.0

# GUI Framework
opencv-python==4.8.1.78
numpy==1.24.3
pyautogui==0.9.54

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# HTTP Requests
requests==2.31.0

# ngrok Integration (Optional but recommended)
pyngrok==7.0.3

# System Tray Support (Optional but recommended for compiled apps)
pystray==0.19.5

# For EXE Compilation
PyInstaller==6.3.0
